fix new expiry price bug/build

+++ programs/drift/src/math/amm/tests.rs
@@ -188,7 +188,7 @@ fn calculate_expiry_price_long_imbalance_with_loss_test() {
     expiry_price = calculate_expiry_price(
         &market.amm,
         oracle_price_data.price,
-        111_111_110 * QUOTE_PRECISION,
+        (111_111_110 * QUOTE_PRECISION) as i128,
     )
     .unwrap();
 
@@ -270,7 +270,7 @@ fn calculate_expiry_price_long_imbalance_test() {
     expiry_price = calculate_expiry_price(
         &market.amm,
         oracle_price_data.price,
-        111_111_110 * QUOTE_PRECISION,
+        (111_111_110 * QUOTE_PRECISION)  as i128,
     )
     .unwrap();
 
@@ -387,7 +387,7 @@ fn calculate_expiry_price_test() {
     expiry_price = calculate_expiry_price(
         &market.amm,
         oracle_price_data.price,
-        111111110 * QUOTE_PRECISION,
+        (111111110 * QUOTE_PRECISION) as i128,
     )
     .unwrap();
